Fix ctx.getImageData
Summary:
# Android

Each mobile device has it's own pixel density or scale (px to dpi ratio). We compensate for this in the canvas to make sure the canvas output looks the same (size/dimension) on each device. Some devices can have floating point pixel densities (e.g., `2.75`), which leads to rational numbers like `22.5` (instead of whole numbers or integers). Depending on the drawing canvas implementation it will floor or round leading to images where pixels in-between might get interpolated (screenshot 1).

When calling the canvas `getImageData` function, it returns the image data in the canvas size not adjusted for pixel density or scale and therefore can be of different size depending on the device. This is important and was implemented that way to make sure image data doesn't loose resolution when using `ctx.putImageData(imageData, x, y)`, otherwise the rendered/put image data will look pixelated on devices with a pixel density `> 1` (screenshot 2). In order to get an image from the image data, PTL has an `ImageUtil.fromImageData(imageData)` function, which will return an image based on the image data size, which means it will have different sizes based on the devices; pixel density (like the one in the first image with interpolated values).

We can somewhat fix both by adjusting the image size when calling `ImageUtil.fromImageData` where it returns an image in the size requested by the `ctx.getImageData(sx, sy, sw, sh)` (screenshot 3) and when using `ctx.putImageData` it will use the image data that respects the device pixel density (screenshot 4).

Reviewed By: chrisklaiber

Differential Revision: D36292404

fbshipit-source-id: e33e50e448e8abf6833b88845c1b1158b872edbe