exclude cxx/test/build folder from yarn bootstrap
Summary:
running "yarn" or "yarn bootstrap" would failed because it is also building the .ts file under cxx/test/build. This diff exclude them

[Here is a helpful link](https://unional.gitbooks.io/typescript/content/pages/06-files-and-projects/tsconfig.html#:~:text=Guidelines-,Build%20config,-Use%20tsconfig.build) to learn about tsconfig.json and tsconfig.build.json.

The tsconfig.build.json is used when `yarn` is called. and tsconfig.json is more for the purpose of configuring IDE -- that's why we want to separate them.

With this config,
1. when building the library, we exclude the test dependencies files and the example app

2.when configuring the IDE, we only exclude the test dependencies (since they are auto genreated files)

Reviewed By: chrisklaiber, ansonsyfang

Differential Revision: D36954496

fbshipit-source-id: 54a3dc15e288b794cf44961cedb140f7fcca75c5