Fix targetResolution setup on our examples
Summary:
We are seeing a crash on flipping camera on AnimeGAN and ObjectDetection examples on iPhone XS.

It turns out that on iPhone XS, back camera supports 4K while front camera only supports 1080p. Default we set the `sessionPreset` to 4K `.hd4K3840x2160` based on the `targetResolution={{width: 1920, height: 1080}}`.

```
} else if width <= 1080 && height <= 1920 {
    captureSession.sessionPreset = .hd1920x1080
} else {
    captureSession.sessionPreset = .hd4K3840x2160
}
```
https://github.com/pytorch/live/blob/2ae0b76a07271b1dba66ff654ad438c313f0b55d/react-native-pytorch-core/ios/Camera/CameraView.swift#L107-L111

This diff fixes the issue by correctly setting the `targetResolution={{width: 1080, height: 1920}}`, so the default `sessionPreset` would be `.hd1920x1080` as expected now.

Follow up:
A better solution would be gracefully handle the `sessionPreset` based on both `targetResolution` and the max support resolution of the current camera.

Reviewed By: raedle

Differential Revision: D37050514

fbshipit-source-id: 8c2191477517b74f9418e2f120dfee57a4d22275