Handle Exception thrown in async task
Summary:
When mobile module is passed with unexpected input and call with async, the app crash. This is easy to verify with Expo with the following testing codes.

The reason is we are running two threads when calling
```
await model.forward()
```
three functions are hosted in the two threads
|UI Thread| Worker Thread|
|SetupFunction|WorkFunction|
|ResolveFunction||

The crash happens because when the setup function (as part of a Promise) throws, it will **NOT** end the thread, but enter the "catch block" -- if the within the catch blocks we do nothing, the thread would go on to start the workFunc.

for example, test the following codesnipt on your browser
```
await new Promise(function(resolve, reject) {
    console.log('promise1');
    throw "something";
}).then(function() {console.log('promise2');}).catch(function(){console.log('promise rejected')});
```

**caveat** calling model with unexpected input would still lead to a crash --
```
let output = await model.forward(1)
```
probably because we are mapping one IValue Type to another -- which destroy the worker thread with a non std::exception. (assert / segmentation fault) etc.

This would be alleviate when we typecheck the inspect type in a follow-up diff

Reviewed By: chrisklaiber

Differential Revision: D37238397

fbshipit-source-id: ed5a969feecebe13cb11181d103e0f26729d71e9