Add flora landing page
Summary:
The purpose of this diff is to add a WIP landing page for Flora so that we can link to it from the blog posts announcing PlayTorch.

As a follow up we will need to add links to PlayTorch in the iOS and Android app store, and a call-to-action at the bottom (perhaps "Join our discord").

Reviewed By: clarksandholtz

Differential Revision: D37540835

fbshipit-source-id: b2899db4d3f193675282eeaf2d42e03fe3ba6791