Update Podfile.lock for templates
Summary:
The `Podfile.lock` in the templates pins LibTorch-Lite to version `1.11.0`. After the recent upgrade to PyTorch Mobile releases version 1.12, the template `Podfile.lock` need to be updated as well.

The GitHub action that tests the `react-native-template-pytorch-live` template fails for that reasons:

https://github.com/pytorch/live/runs/7134251602?check_suite_focus=true

Updating `Podfile.lock` for both templates to use `LibTorch-Lite` `1.12.0`.

Reviewed By: chrisklaiber, clarksandholtz

Differential Revision: D37557339

fbshipit-source-id: 52963ef1045817ef246055863b1c0f042580ef72