enable generic dict as module input
Summary:
As title, we enable the dictionary as input. Since dictionary is a common format of model output, this feature will enable a more generic API of chaining of model)

Torchscript doesn't support recursive Dict structure as input (i.e dict[string, [dict[str]] ) at this moment, so it suffices to test it as a top-level structure.

Reviewed By: ansonsyfang

Differential Revision: D37574283

fbshipit-source-id: ffa1a5baa9e0b9d4d7812c93acea6191c02e5f1b