Refactor JSI Playground for Easier E2E Testing
Summary:
As the JSI library becoming larger and larger, the old UI (with one button) become not efficient -- it only result to a long wait-time for unit testing, but also generate a lot of redundant output in console.

In this diff, we redesign the Testing UI for easier in the future.

To test new function, simply add a new item to the `testUnitList` with the following fields
```
{
name: ... // name of the tests
testFunc: ... // functions you want to test
}
```

Reviewed By: raedle

Differential Revision: D37372776

fbshipit-source-id: 069d2bb266ad8e93512e9ec1be59a898ef13a2e0