Add PlayTorch branded landing page with app store links
Summary:
This adds a snack landing page at `[playtorch base url]/snack/[optional snack id]`.

`../expo/...` can also be used.

If the user goes to `.../snack/` with no identifier it redirects them to the starter snack. If the id is present it shows a landing page with options to open in playtorch or download the app.

In an upcoming diff I'll use Clark's app store buttons and conditionally show them if the user is on ios/android.

Reviewed By: raedle

Differential Revision: D37900585

fbshipit-source-id: 61941d8df12287ac38cdbb3343420d8e0f4852c8