Add example landing page
Summary:
This adds a simple example landing page for examples. In the next diff I'll add the example title and validation for the slugs.

FWIW this is broken in the production build due to the same baseUrl issue that raedle encountered. Fortunately that won't be an issue as we are setting our baseUrl to "/"

Reviewed By: clarksandholtz

Differential Revision: D37968850

fbshipit-source-id: 2f14795a47352abce15e6b08f372deda56c7a7f0