Replace url in history rather than pushing for example url with no example slug
Summary:
When the user navigates to /example/ with no slug we show them a Page Not Found. Right now that adds a page to the browser history, so when they navigate back it jumps them forward again via the redirect.

This diff changes that to simply replace the `/example/` entry, so the user can now navigate back to where they were before.

Reviewed By: raedle

Differential Revision: D37977066

fbshipit-source-id: 86275fc8bfc891ec98e1e51aa535ae644919ffc7