Fix Github action workflow build-template-ios (#80)
Summary:
This diff tries to fix build-template-ios workflow error.
The problem seems to be mixing objc and cdecl in Swift and import the generated header in a file that uses both these interop functions.
This diff converts the cdecl functions to objc functions. It fixes build-template-ios on my test: https://github.com/ansonsyfang/live/actions/runs/2744959171
I'm not too sure about the root cause, my hunch is something related to name mangling. Apparently it's been fixed in Xcode 13.3 and above.

Pull Request resolved: https://github.com/facebookresearch/playtorch/pull/80

Reviewed By: liuyinglao

Differential Revision: D38196171

fbshipit-source-id: 6c1f7ba5486a310e4de6e7395941c56adf801a91