Fix hydration error
Summary:
There's a hydration error that occurs when the result of `isLikelyIOSDevice` is different on the client and the server, which causes the site to render the "Join the iOS Beta" button with the **google play store link**.

This fixes it by preventing the app store buttons from rendering at all on the server.

We didn't catch this in development because the docusaurus dev server doesn't do the same server rendering.

Reviewed By: raedle

Differential Revision: D38378259

fbshipit-source-id: 314ae6f62c1c3db7ca0a1ddce70d21e59411e576