Fix mismatch and add comment for paramCount in setPropertyHostFunction
Summary:
# Context
The `paramCount` parameter in `setPropertyHostFunction` is used to set the [function.length](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/length) property in JSI.

The property indicates the number of parameters expected by the function. In our case, this is the minimal number of required arguments of the function.

# This Diff

There has been a few mismatches between `paramCount` and the minimal arg count in older code. This is a minor issue as `function.length` is only reflected in the metadata and doesn't affect function execution, but let's still fix these mismatches and add comments for the parameter to avoid future mistakes.

Reviewed By: liuyinglao

Differential Revision: D38294191

fbshipit-source-id: e1fa7788c4b3bf087561f40ee3bc7181b4a64b1e