convert torchlive.torch namespace to jsi::Object
Summary:
# Context

The idea is that things with state are "HostObjects" (like Tensor, Blob, etc), and those without state are "namespaces" (like media, torch, torch.jit, etc).

This diff replaces `TorchHostObject` with `TorchNamespace`. After the previous refactoring diffs, `TorchHostObject` is now just a collection of values and is ready to be converted to a namespace object.

**A minor update**
Removed all comment blocks before functions since we already have everything in the JS-facing docs. Keeping the documentation in two places can easily get out of sync.

Reviewed By: chrisklaiber

Differential Revision: D38449826

fbshipit-source-id: b8e1fc6a8aceea2df8e01d0e431c19bb057de6c2