Properly handle canvas invalidate method clearing or not clearing previous drawings
Summary:
Proper behaviors of `invalidate` and `clear`:

`invalidate` - apply the the staged drawings on canvas, and clear the staged drawings.
`clear` - clear the existing drawings on canvas, and clear the staged drawings.

Based on these behaviors, the following code should handle them nicely.
```
    func invalidate() {
        let sublayers = self.sublayers
        self.sublayers.removeAll()

        DispatchQueue.main.async {
            self.renderLayers(allLayerData: sublayers, layer: self.layer)
            self.layer.layoutIfNeeded()
        }
    }

    func clear() {
        sublayers.removeAll()

        DispatchQueue.main.async {
            self.layer.sublayers?.removeAll()
            self.layer.layoutIfNeeded()
        }
    }
```
Yet, if you test it on the Canvas Animation example, your eyes hurt 🤣
https://pxl.cl/29xPK

One hypothesis is we are not batching the drawings over JS and native, and it causes bridge congestion.
Testing it with less drawings, it doesn't get better.
https://pxl.cl/29xPN

The issue is we switch between main thread and JS background thread too often in Canvas Animation example.
Instead of clearing existing drawings on canvas right away on `clear` called, we defer it and handle it in `invalidate`.
This reduces the thread switching and the animation now works.

Reviewed By: raedle

Differential Revision: D38481896

fbshipit-source-id: 40c9d3181ab777b98c9547bc0d778294c71c4f92