Lazy install JSI
Summary:
To install the PlayTorch JSI API and expose bindings on the JSI Runtime global scope, the Android app needs to implement `protected JSIModulePackage getJSIModulePackage()` and return the JSI module.

While this works, it requires manual effort by devs to set it up for the Android app. This also doesn't work with Expo managed apps and requires devs to eject Expo apps.

This change installs the PlayTorch JSI API lazily and works with auto-linking.

Reviewed By: chrisklaiber

Differential Revision: D38687374

fbshipit-source-id: 9bf801b7d3565032247e89e1c84726ca28d03348