Match any exception type for Tensor indexing API
Summary:
The tensor indexing API relies on exceptions when a prop name cannot be converted to an int via `std::stoi`.

After upgrading to Expo 46, it causes a runtime error (`std::exception&`), which doesn't get caught with the current `catch`. The change broadens the `catch` to match any exception (`catch (...)`), which fixes the issue.

The reasons are not clear, but it's assumed to be a compiler flag change. We continue to investigate, but in the meantime, this change unblock anyone who is facing the same issue

Reviewed By: justinhaaheim

Differential Revision: D39148492

fbshipit-source-id: 766d716d7991415110262d600d94268441e858af