Match any exception type for Tensor indexing API (#118)
Summary:
Pull Request resolved: https://github.com/facebookresearch/playtorch/pull/118

The tensor indexing API relies on exceptions when a prop name cannot be converted to an int via `std::stoi`.

After upgrading to Expo 46, it causes a runtime error (`std::exception&`), which doesn't get caught with the current `catch`. The change broadens the `catch` to match any exception (`catch (...)`), which fixes the issue.

The reasons are not clear, but it's assumed to be a compiler flag change. We continue to investigate, but in the meantime, this change unblock anyone who is facing the same issue

Reviewed By: justinhaaheim

Differential Revision: D39148492

fbshipit-source-id: f52331671e57ce9957befa3717c43b6f7b7de76c