Prevent OOM for MNIST examples
Summary:
The MNIST example in the PlayTorch app crashes after drawing several numbers on the canvas. The canvas context2d needs to be cleared after drawing.

The change clears the context2d before drawing on the canvas to avoid OOM.

Also removing the `deprecated` directive from `clear` while it is needed to certain canvas use cases

Reviewed By: liuyinglao

Differential Revision: D39418643

fbshipit-source-id: 28f1662c38560bbb289eeb224fdaf951bc0b95e9