Added support for ops where self is not first argument
Summary:
Added support for ops where self is not first argument

instead of `op.arguments[1:]`, this version creates a `get_arguments_list()` function that will return a copy of the arguments list without self.

A copy is made because we will eventually be using this code to generate `TorchNamespace.cpp` which will need access to self.

Differential Revision: D41082852

fbshipit-source-id: e6d41fcb6973c2f8bb98795a28be331836daaacd