Fix issue with JNI Env not available for audio
Summary: The Hermes GC will eventually clean up orphan host objects. This is also the case for `AudioHostObject`, which itself has a reference to `Audio` that needs to be cleaned up, hence the call to the `Audio::~Audio` destructor. On Android, it will try to release the `global_ref<JAudio>` resource, which at the time, may not run in the JVM thread scope causing an app crash.

Differential Revision: D42060942

fbshipit-source-id: 02b26b4691968c6b172df1c01525a026077fd837