Fix issue with JNI Env not available for audio (#182)
Summary:
Pull Request resolved: https://github.com/facebookresearch/playtorch/pull/182

The Hermes GC will eventually clean up orphan host objects. This is also the case for `AudioHostObject`, which itself has a reference to `Audio` that needs to be cleaned up, hence the call to the `Audio::~Audio` destructor. On Android, it will try to release the `global_ref<JAudio>` resource, which at the time, may not run in the JVM thread scope causing an app crash.

Reviewed By: ansonsyfang

Differential Revision: D42060942

fbshipit-source-id: b19f723f361db3794b704be97456bd4bbe980848