Workaround
Modified the file PyTorchCoreJSI.mm to repeatedly try installing
\_\_torchlive\_\_ if RCTCxxBridge.runtime is null, rather than
just giving up.

Modfied the file install-jsi.ts to run all of it's side-effect logic in
a function that does a "return;" if the variable "PlayTorchJSIModule" is null,
instead of throwing an error. Also added an extra side effect that adds
a function to the global object, that can be used to await the existence
of \_\_torchlive\_\_, before calling other PyTorch code.